Extract component to separate file
